+++ spec/unit/hanami/router/inspector_spec.rb
@@ -3,44 +3,48 @@
 require "hanami/router/inspector"
 
 RSpec.describe Hanami::Router::Inspector do
-  subject { described_class.new(routes: routes) }
-  let(:routes) { [] }
-
-  describe "#initialize" do
-    it "returns a frozen instance of #{described_class}" do
-      expect(subject).to be_kind_of(described_class)
-    end
-  end
-
   describe "#add_route" do
     it "adds a route to the inspector" do
-      subject.add_route(Hanami::Router::Route.new(http_method: "GET", path: "/", to: "home#index"))
-      expect(subject.call).to include("GET     /                             home#index")
+      inspector = described_class.new
+
+      inspector.add_route(Hanami::Router::Route.new(http_method: "GET", path: "/", to: "home#index"))
+
+      expect(inspector.call).to include("home#index")
     end
   end
 
   describe "#call" do
-    context "no routes" do
-      it "returns an empty result" do
-        expect(subject.call(routes)).to eq("")
-      end
+    it "forwards to the formatter with the given routes" do
+      routes = [
+        Hanami::Router::Route.new(http_method: "GET", path: "/", to: "home#index", as: :root, constraints: {}),
+        Hanami::Router::Route.new(http_method: "GET", path: "/about", to: "home#about", as: :about, constraints: {})
+      ]
+      formatter = ->(rs) { rs.map(&:path).join("+") }
+
+      inspector = described_class.new(routes: routes, formatter: formatter)
+
+      expect(inspector.call).to eq("/+/about")
     end
 
-    context "with routes" do
-      let(:routes) do
-        [Hanami::Router::Route.new(http_method: "GET", path: "/", to: "home#index", as: :root, constraints: {}, blk: nil)]
-      end
+    it "forwards arguments to the formatter" do
+      routes = [
+        Hanami::Router::Route.new(http_method: "GET", path: "/", to: "home#index", as: :root, constraints: {}),
+        Hanami::Router::Route.new(http_method: "GET", path: "/about", to: "home#about", as: :about, constraints: {})
+      ]
+      formatter = ->(rs, join_with:) { rs.map(&:path).join(join_with) }
+      inspector = described_class.new(routes: routes, formatter: formatter)
 
-      it "returns inspected routes" do
-        expected = [
-          "GET     /                             home#index                    as :root"
-        ]
+      expect(inspector.call(join_with: "-")).to eq("/-/about")
+    end
 
-        actual = subject.call
-        expected.each do |route|
-          expect(actual).to include(route)
-        end
-      end
+    it "defaults to the human friendly formatter" do
+      routes = [
+        Hanami::Router::Route.new(http_method: "GET", path: "/", to: "home#index", as: :root, constraints: {})
+      ]
+
+      inspector = described_class.new(routes: routes)
+
+      expect(inspector.call).to eq(Hanami::Router::Formatter::HumanFriendly.new.call(routes))
     end
   end
 end